Add support for short datetimes
This commit adds support for short datetimes.  Short datetimes allow the
user to specify a datetime with a much shorter and concise syntax which
acts relatively to the current date.  It consists of a delta and a
modifier.  A delta can be either an integer, representing that may days
from the current day, or the character '.' which represents the current
day.  If a delta is suffixed with the modifier '^' then the time of the
given day is set to 00:00.  If it is suffixed with the modifier '$' then
the time is set to 23:59.