Add the `addtask' utility
The `addtask' utility allows you to conveniently add tasks to a task
directory in an ordered fashion.  It accepts files as arguements the
same way utilities like `cat' do, including special handling of the `-'
file argument.  The utility still requires a manual page, but afterwards
a new version of the taskutils collection will be ready for release.