Save tasks by task title
Instead of saving tasks with numerical IDs, we now save them with the
filename being the task title.  This helps with several logistical
issues such as someone removing a file, and makes it easier for the user
to try to find the correct task file as they can just check the
filenames.